attempt 2 to fix pylint errors by creating file of common test functions moved common functions to test folder for ease of access
