+++ persimmon/website/tests/test_bank_account.py
@@ -1,6 +1,6 @@
 from django.test import TestCase, Client
 from django.urls import reverse
-from ..common import make_user
+from common_test_functions import make_user
 
 from ..models import EmployeeLevel, BankAccount, AccountType, ApprovalStatus
 from .. import views
