+++ persimmon/website/tests/test_user_banktransfer.py
@@ -1,6 +1,6 @@
 from django.test import TestCase, Client
 from django.urls import reverse
-from ..models import User, EmployeeLevel, BankAccount, AccountType, BankStatements
+from ..models import EmployeeLevel, BankAccount, AccountType, BankStatements
 from .. import views
 from .common_test_functions import make_user, view_pending_account, approve_account
